# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for personal finance management with AI features.",
        "title": "Expense Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8989",
    "basePath": "/api/v1",
    "paths": {
...
        "/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SummaryResponse"
                        }
                    }
                }
            }
        },
        "/summary/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get category summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CategorySummaryResponse"
                            }
                        }
                    }
                }
            }
        },
        "/summary/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get monthly summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month YYYY-MM",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MonthlySummaryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategorySummaryResponse": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "dto.MonthlySummaryResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                },
                "total_expense": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                }
            }
        },
...
        "dto.SummaryResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "total_expense": {
                    "type": "number"
                },
                "total_income": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}